minimise sending reads to stdout

+++ modules/VertRes/Pipelines/Import_iRODS_fastq.pm
@@ -252,16 +252,31 @@ for my \$fastq (\@fastqs)
 }
 
 # output reads with PF pass only
-system("samtools view -F 0x200 $in_bam > pf_pass.sam");
+system("samtools view -F 0x200 -b -o pf_pass.bam $in_bam");
 
 # Reads with PF fail have bases set to N and quality scores to 0.
 system("samtools view -f 0x200 $in_bam | awk -F '\\t'  'BEGIN{OFS=\\"\\t\\";} {gsub(/[ACGT]/,\\"N\\",\\\$10) }; {gsub(/./,\\"!\\",\\\$11) };   1' > pf_fail.sam");
 
-#  remove secondary alignments
-system("samtools view -H $in_bam | cat - pf_pass.sam pf_fail.sam | samtools view -F 0x100 -b -S - > secondary_alignments_removed.bam");
+if(-s qq[pf_fail.sam] == 0)
+{
+  system("samtools view -F 0x100 -b -o secondary_alignments_removed.bam pf_pass.bam");
+}
+else
+{
+  system("samtools view -H $in_bam > header.sam");
+  system("cat header.sam pf_fail.sam | samtools view -b -S  > pf_fail.bam");
+  unlink("pf_fail.sam");
+  unlink("header.sam");
+  
+  system("samtools merge -n merged.bam pf_pass.bam pf_fail.bam");
+  unlink("pf_pass.bam");
+  unlink("pf_fail.bam");
+  
+  system("samtools view -F 0x100 -b -o secondary_alignments_removed.bam  merged.bam");
+  unlink("merged.bam");
+}
+
 system("mv secondary_alignments_removed.bam $in_bam");
-unlink("pf_pass.sam");
-unlink("pf_fail.sam");
 
 VertRes::Wrapper::samtools->new()->sort(qq[$in_bam], qq[sorted], n => 1, m => $samtools_sorting_memory);
 system("mv sorted.bam $in_bam");
